refactor: AuthService의 위치 /auth 추가
